% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{add_fields_to_table}
\alias{add_fields_to_table}
\title{向 SQLite 数据库表中添加新字段}
\usage{
add_fields_to_table(db_path, table_name, new_field_names)
}
\arguments{
\item{db_path}{一个字符型字符串，指定 SQLite 数据库的文件路径。}

\item{table_name}{一个字符型字符串，指定要修改的数据库表名。}

\item{new_field_names}{一个字符型向量，包含要添加的新字段名称。}
}
\value{
一个列表，其中包含两个字符型向量：`success` 和 `failed`。`success` 向量包含成功添加的字段名称，`failed` 向量包含失败的字段名称及其失败原因。
}
\description{
这个函数将指定的新字段添加到 SQLite 数据库中的指定表。如果字段已经存在，则记录这些字段为失败。所有新字段的类型被设置为 TEXT。
}
\examples{
db_path <- "path/to/your/database.sqlite"
table_name <- "your_table"
new_field_names <- c("new_column1", "new_column2")
result <- add_fields_to_table(db_path, table_name, new_field_names)
print(result)

}
