% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advance_breeding_stage.R
\name{advance_breeding_stage}
\alias{advance_breeding_stage}
\title{处理初级产比进入高级产比的数据}
\usage{
advance_breeding_stage(
  candidates,
  start_num = 1,
  selection_key = "高级产比",
  target_key = "国区省区"
)
}
\arguments{
\item{candidates}{数据框,要准备晋级的数据}

\item{start_num}{整数，起始编号，默认为1}
}
\value{
数据框，处理后的高级产比数据
}
\description{
只有初级产比进入高级产比时，世代不增加
}
