% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.R
\name{calculate_muchan_increase_with_multiple_methods}
\alias{calculate_muchan_increase_with_multiple_methods}
\title{计算增产百分比并返回排序结果}
\usage{
calculate_muchan_increase_with_multiple_methods(df)
}
\arguments{
\item{df}{数据框，必须包含以下列：
\itemize{
  \item muchan: 产量数据
  \item is_ck: 是否为对照组，值为1表示对照组，其他值表示实验组
}}
}
\value{
返回数据框，新增以下列：
  \itemize{
    \item 较临近对照增产%: 与最近两个对照点的增产百分比
    \item 较平均对照增产%: 与所有非空对照的增产百分比
    \item 较临近对照位次: 根据增产百分比与最近两个对照点比较的位次
    \item 较平均对照位次: 根据增产百分比与所有非空对照比较的位次
  }
}
\description{
该函数根据两种方式计算增产百分比：一种是与最近两个对照点比较，另一种是与所有非空对照的平均值比较。函数还会返回每种方式的增产百分比，并为每种方式计算位次。
}
\examples{
df <- data.frame(
  muchan = c(10, 20, 15, 25, 30, 10),
  is_ck = c(1, 0, 1, 0, 0, 1)
)
result <- calculate_muchan_increase_with_multiple_methods(df)
print(result)

}
