% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination.R
\name{combination}
\alias{combination}
\title{杂交组合}
\usage{
combination(
  ma = c("JD12", "JD17"),
  pa = c("ZLD6001", "ZLD6024", "ZLD6033"),
  memo = NA
)
}
\arguments{
\item{ma}{向量，母本列表.}

\item{pa}{向量，父本列表}

\item{memo}{字符串， 备注}
}
\value{
两组亲本所配制的组合
}
\description{
根据父母本配制杂交组合
}
\examples{
combination(ma=c("冀豆12", "冀豆17"),pa=c("中联豆6001", "中联豆6024", "中联豆6033"),memo = "转基因")
combination(ma=c("冀豆12", "冀豆17"),pa=c("中联豆6001", "中联豆6024", "中联豆6033"),memo = NA)
}
