% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination.R
\name{combination_fromfile}
\alias{combination_fromfile}
\title{文件中杂交组合}
\usage{
combination_fromfile(filename, only = TRUE, order = FALSE)
}
\arguments{
\item{filename}{文件名，里面包括3列，顺序为为母本，父本和备注，其它列无要求，文件格式为xlsx}

\item{only}{逻辑值，是否去除重复组合}

\item{order}{是否配制的组合进行排序}

\item{prefix}{组合的前缀}
}
\value{
返回从文件中配制的杂交合组
}
\description{
文件中杂交组合
}
