% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination.R
\name{combination_matrix}
\alias{combination_matrix}
\title{获得组合矩阵}
\usage{
combination_matrix(my_combi)
}
\arguments{
\item{my_combi}{data.frame, 里面要有三列数据:分别为 ma, pa, stageid}
}
\value{
返回组合矩阵
}
\description{
获得组合矩阵
}
\examples{
combination_matrix(data.frame(ma=c("JD12","JD17","JD32"),pa=c("ZLD6001","ZLD6003","ZLD6024"),stageid=c("ZJ001","ZJ002","ZJ003")))
}
