% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_trait.R
\name{create_class_matrix}
\alias{create_class_matrix}
\title{根据 fieldid 和 xingzhuang 组合生成 class 展示矩阵}
\usage{
create_class_matrix(df)
}
\arguments{
\item{df}{一个数据框，必须包含以下列：fieldid, xingzhuang 和 class。}
}
\value{
返回一个宽格式的数据框，以 fieldid 和 xingzhuang 为行列，显示 class 值。
}
\description{
这个函数接受一个数据框，并生成一个以 fieldid 和 xingzhuang 为行列的宽格式数据框，
其中包含每个组合的 class 值。如果有重复的 (fieldid, xingzhuang) 组合，将保留最后一个记录。
}
\examples{
df <- data.frame(
  non_fieldid = c("12_2_Excellent", "20_2_Brown", "20_99_Segregating", "36_1_Lanceolate", "36_2_Ovate", "36_4_Round"),
  fieldid = c("f246oa2H01028", "f246oa2H01028", "f246oa2H01028", "f246oa2H01059", "f246oa2H01059", "f246oa2H01059"),
  xingzhuang = c("huaqipingjia", "rongmaose", "rongmaose", "yexing", "yexing", "yexing"),
  class = c("优", "棕", "分离", "披针", "卵圆", "圆"),
  class_C = c("Excellent", "Brown", "Segregating", "Lanceolate", "Ovate", "Round"),
  trait_code = c(12, 20, 20, 36, 36, 36),
  class_code = c(2, 2, 99, 1, 2, 4),
  number = c(19, 51, 52, 88, 89, 91)
)
create_class_matrix(df)
}
