% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_trait.R
\name{create_multiple_pairs_from_df}
\alias{create_multiple_pairs_from_df}
\title{创建配对数据框}
\usage{
create_multiple_pairs_from_df(df)
}
\arguments{
\item{df}{一个包含元素的单列数据框}
}
\value{
一个包含两个变量的数据框，一个变量是以 "f" 开头的元素，另一个变量是不以 "f" 开头的元素
}
\description{
该函数遍历输入数据框中的每个元素，如果元素以 "f" 开头，接着是任意7个字符，然后是5个数字，
则将其存入一个临时变量。如果元素不匹配该模式，则将临时变量和该元素组合，并存入一个新的数据框中。
}
\examples{
elements_df <- data.frame(elements = c("fabcdefg12345", "bar", "baz", "fghijklm67890", "buzz", "buzzie", "fnopqrst54321", "buzzbuzz"), stringsAsFactors = FALSE)
pairs_df <- create_multiple_pairs_from_df(elements_df)
print(pairs_df)
}
