% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_plot.r
\name{designPlot}
\alias{designPlot}
\title{设计地块布局}
\usage{
designPlot(
  y = 10,
  bridges = c(9, rep(6, 10), rep(3, 9)),
  distanceBetweenBlocks1 = 1,
  distanceBetweenBlocks2 = 0.2,
  protectedColumns = c(3),
  protectedBlocks = c(1, 3, 4, 6),
  waterColumns = c(5),
  laneColumns = c(9),
  protectedArea = c(56, 65, 6, 8, 34, 38, 2, 5),
  designFromLeft = TRUE,
  plantFromLeft = TRUE,
  subg = 3
)
}
\arguments{
\item{y}{每块地的列数。默认值为10。}

\item{bridges}{每块地的宽度。默认值为c(9, rep(6, 10), rep(3, 9))。}

\item{distanceBetweenBlocks1}{宽条间距，默认值为1。}

\item{distanceBetweenBlocks2}{窄条间距，默认值为0.2。}

\item{protectedColumns}{保护行所在的列，默认值为c(3)。}

\item{protectedBlocks}{保护行所在的块，默认值为c(1, 3, 4, 6)。}

\item{waterColumns}{水沟所在的列，默认值为c(5)。}

\item{laneColumns}{道路所在的列，默认值为c(9)。}

\item{protectedArea}{植被区域，默认值为c(56, 65, 6, 8, 34, 38, 2, 5)。}

\item{designFromLeft}{是否从左侧设计布局，默认值为TRUE。}

\item{plantFromLeft}{是否从左侧种植，默认值为TRUE。}

\item{subg}{子组，默认值为3。}

\item{blocks}{地块数。默认值为20。}
}
\value{
返回设计的地块布局矩阵。
}
\description{
设计一个包含水沟、道路和保护行的地块布局。
}
