% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planting.R
\name{ensure_and_reorder_columns}
\alias{ensure_and_reorder_columns}
\title{确保数据框包含指定字段并调整字段顺序}
\usage{
ensure_and_reorder_columns(df, required_columns)
}
\arguments{
\item{df}{数据框，需要检查并调整字段的顺序。}

\item{required_columns}{字符串向量，包含所有需要的必需字段名。}
}
\value{
返回调整后的数据框，其中必需字段在前，其他字段在后。
}
\description{
该函数确保数据框中包含指定的必需字段。如果数据框中缺少某个必需字段，则自动添加并填充为 `NA`。函数还将必需字段调整至数据框的前面，其余字段保持在后面。
}
\examples{
# 定义数据框
df <- data.frame(name = c("A", "B"), stageid = c(1, 2))

# 定义必需字段
required_columns <- c("id", "user", "stageid", "name", "ma", "pa", "mapa", "memo",
                      "stage", "next_stage", "f", "sele", "process", "path", "source")

# 调用函数确保字段存在并重新排序
df <- ensure_and_reorder_columns(df, required_columns)

# 查看结果
print(df)

}
