% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_plot.r
\name{expand_rows}
\alias{expand_rows}
\title{根据指定列的值复制数据框行并添加序列号}
\usage{
expand_rows(df, replicate_col = "rows", sequence_col = "seq_id")
}
\arguments{
\item{df}{要扩展的数据框（data.frame）}

\item{replicate_col}{字符型，指定包含复制次数的列名，默认为"rows"}

\item{sequence_col}{字符型，指定新序列号列的列名，默认为"seq_id"}
}
\value{
返回一个新的数据框，其中：
\itemize{
  \item 包含原始数据框的所有列
  \item 新增一个序列号列，显示每组复制中的序号
}
}
\description{
该函数根据数据框中指定列的值复制每一行，并为每组复制添加序列号列。
}
\examples{
# 示例数据
sample_data <- data.frame(
  id = c(1, 2, 3),
  name = c("A", "B", "C"),
  times = c(2, 3, 1)
)

# 使用'times'列扩展行
expanded_data <- expand_rows(sample_data, "times", "copy_num")
print(expanded_data)

# 输出结果：
#   id name times copy_num
# 1  1    A     2        1
# 2  1    A     2        2
# 3  2    B     3        1
# 4  2    B     3        2
# 5  2    B     3        3
# 6  3    C     1        1

}
