% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_id.R
\name{generate_id}
\alias{generate_id}
\title{Generate IDs with specified parameters}
\usage{
generate_id(
  start_num = 1,
  end_num,
  char = "g",
  digit_length = 5,
  include_datetime = TRUE
)
}
\arguments{
\item{start_num}{An integer indicating the starting number of the ID sequence.}

\item{end_num}{An integer indicating the ending number of the ID sequence.}

\item{char}{A character placed at the beginning of each ID.}

\item{digit_length}{An integer specifying the length of the numeric part, controlling the padding.}

\item{include_datetime}{A logical value indicating whether to include the current year, month, day, hour, minute, and second in the ID.}
}
\value{
A vector containing all the generated IDs between start_num and end_num.
}
\description{
This function generates a series of IDs based on the specified parameters.
}
\examples{
\dontrun{
start <- 123
end <- 135
character <- "A"
digit_length <- 5
result_ids <- generate_id(start, end, character, digit_length, include_datetime = TRUE)
for (result_id in result_ids) {
  print(paste("Resulting ID:", result_id))
}
}

}
