% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_productivity.R
\name{get_advanced}
\alias{get_advanced}
\title{处理初级产比进入高级产比的数据}
\usage{
get_advanced(my_primary, start_num = 1)
}
\arguments{
\item{my_primary}{数据框，包含初级产比信息}

\item{start_num}{整数，起始编号，默认为1}
}
\value{
数据框，处理后的高级产比数据
}
\description{
只有初级产比进入高级产比时，世代不增加
}
