% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination.R
\name{get_combination}
\alias{get_combination}
\title{获得组合列表}
\usage{
get_combination(
  ma = c("JD12", "JD17", "JD32"),
  pa = c("ZJD6001", "ZJD6019", "ZJD6024"),
  memo = NA,
  prefix = "ZJ",
  startN = 1
)
}
\arguments{
\item{ma}{向量，母本列表}

\item{pa}{向量，父本列表}

\item{memo}{字符串， 备注}

\item{prefix}{组合前缀}

\item{startN}{起始编号}
}
\value{
返回杂交列表
}
\description{
根据父母本获得组合列表
}
\examples{
get_combination(ma=c("JD12", "JD17"),pa=c("ZLD6001", "ZLD6024", "ZLD6033"),memo = "transgene",prefix="ZJ",startN=1)
get_combination(ma=c("JD12", "JD17"),pa=c("ZLD6001", "ZLD6024", "ZLD6033"),prefix="ZJ",startN=101)
}
