% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination.R
\name{get_combination_fromfile}
\alias{get_combination_fromfile}
\title{配置杂交组合}
\usage{
get_combination_fromfile(filename, prefix = "ZJ24")
}
\arguments{
\item{filename}{文件名，里面包括3列，分别为母本，父本和备注，其它列无要求，文件格式为xlsx.}

\item{prefix}{组合的前缀。}
}
\description{
配置杂交组合
}
