% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{get_line}
\alias{get_line}
\title{处理单株进入株行的数据}
\usage{
get_line(my_plant, start_num = 1)
}
\arguments{
\item{my_plant}{数据框，必需含有name,next_stage,f字段}

\item{start_num}{整数，起始编号，默认为1}
}
\value{
数据框，处理后的株行数据
}
\description{
只有单株进入株行时，世代不增加
}
