% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_id.R
\name{get_next_id}
\alias{get_next_id}
\title{Get the next ID from a primary ID}
\usage{
get_next_id(my_primary)
}
\arguments{
\item{my_primary}{A data frame containing primary IDs.}
}
\value{
A list containing the next numeric part of the ID and the prefix part of the ID.
}
\description{
This function extracts the current ID and returns the next ID.
}
\examples{
\dontrun{
# Example usage:
my_primary <- data.frame(ID = c("g24040412094300123", "g24040412094300124"))
result <- get_next_id(my_primary)
}

}
