% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{get_population}
\alias{get_population}
\title{升级群体}
\usage{
get_population(my_combi)
}
\arguments{
\item{my_combi}{数据框，从get_combination或get_population函数所产生数据}
}
\value{
下阶段值为“群体”记录升级为群体
}
\description{
从杂交组合中或群体中升级群体
}
\examples{
get_population(my_combi)

}
