% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_plot.r
\name{getSta}
\alias{getSta}
\title{获取地块的统计信息}
\usage{
getSta(fi = designPlot())
}
\arguments{
\item{fi}{设计布局矩阵，默认使用 designPlot 函数生成的矩阵。}
}
\value{
返回统计信息，包括每个地块的名称、道路总数、平均道路宽度以及总数。
}
\description{
对设计布局进行统计分析，计算每个地块的相关指标。
}
\examples{
fi <- designPlot()
stats <- getSta(fi)
print(stats)
}
