% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_plot.r
\name{getSub}
\alias{getSub}
\title{获取子组}
\usage{
getSub(j = 1, vec = c(0, 0, -77, 0, 0, 0), subg = 3)
}
\arguments{
\item{j}{子组的起始索引。}

\item{vec}{输入的向量。}

\item{subg}{子组的大小。}
}
\value{
返回提取的子组。
}
\description{
从向量中提取子组。
}
\examples{
vec <- c(0, 0, 0, -77, 0, 0, 0, 0, -77, 0, 0, 0, 0, 0, 0, 0, -9, 0, 0)
sub_vec <- getSub(3, vec, 3)
print(sub_vec)
}
