% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{increment_generation_name}
\alias{increment_generation_name}
\title{增加名称中的世代数字}
\usage{
increment_generation_name(names, f_vals, max_gen = 9)
}
\arguments{
\item{names}{字符向量，材料名称}

\item{f_vals}{数值向量，对应的世代数}

\item{max_gen}{整数，最大世代数，超过则不再追加，默认为9}
}
\value{
字符向量，处理后的名称
}
\description{
当名称末尾已有冒号分隔的世代号时，替换它；否则追加新的世代号。
}
\examples{
increment_generation_name(c("冀豆12:3", "冀豆15"), c(3, 4))

}
