% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_ck_rows}
\alias{insert_ck_rows}
\title{向数据框中插入对照行}
\usage{
insert_ck_rows(df, ck, first_as_ck = FALSE)
}
\arguments{
\item{df}{数据框，分组后的数据}

\item{ck}{字符向量，对照名称}
}
\value{
在每个分组末尾追加对照行的数据框
}
\description{
在每个分组中插入对照行（NA行），用于田间试验设计。
}
