% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_plot.r
\name{makelist}
\alias{makelist}
\title{创建种子列表}
\usage{
makelist(
  myd = data.frame(num = 1:3, name = c("JD12", "JD17", "五星1"), rp = c(3, 10, 6))
)
}
\arguments{
\item{myd}{输入的数据框，默认包含三列（编号、名称、重复次数）。}
}
\value{
返回一个数据框，其中包含编号和重复元素的列表。
}
\description{
根据输入的数据框，生成包含指定重复次数的元素的列表。
}
\examples{
data <- data.frame(num = 1:3, name = c('JD12', 'JD17', '五星1'), re = c(3, 10, 6))
result <- makelist(data)
print(result)
}
