% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_plot.r
\name{plant}
\alias{plant}
\title{在设计图中种植种子}
\usage{
plant(fi = designPlot(), seed = makelist(), returnfi = TRUE)
}
\arguments{
\item{fi}{矩阵，设计图矩阵，默认使用 \code{designPlot()} 生成。}

\item{seed}{矩阵，种子列表，默认使用 \code{makelist()} 生成。}
}
\value{
矩阵，返回处理后的设计图矩阵。
}
\description{
该函数用于在设计图中根据种子列表进行种植。根据种子列表中的信息，
将设计图中相应的位置替换为种子的详细信息。
}
\examples{
design <- matrix(1:12, nrow = 3)
seeds <- matrix(c("TypeA", "Seed1", "TypeB", "Seed2"), ncol = 2)
colnames(design) <- c("V1", "V2", "V3", "V4")
plant(design, seeds)
}
