% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_plot.r
\name{plantByGroup}
\alias{plantByGroup}
\title{根据子组标记矩阵}
\usage{
plantByGroup(
  ma = matrix(rep(0, 200), nrow = 10, ncol = 20),
  subg = 3,
  mysign = -6
)
}
\arguments{
\item{ma}{输入矩阵。}

\item{subg}{子组的大小。默认值为3。}

\item{mysign}{用于标记的符号。默认值为-6。}
}
\value{
返回标记后的矩阵。
}
\description{
将输入矩阵按行分组，并根据条件对组内元素进行标记。
}
\examples{
ma <- matrix(rep(0, 200), nrow = 10, ncol = 20)
ma[, 3] <- -77
ma[, 18] <- -77
result <- plantByGroup(ma, subg = 6)
print(result)
}
