% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_trait.R
\name{read_and_process_planting_tables}
\alias{read_and_process_planting_tables}
\title{读取文件夹中所有 Excel 文件的 planting 表，并处理字段}
\usage{
read_and_process_planting_tables(folder_path)
}
\arguments{
\item{folder_path}{字符串，表示文件夹的路径。}
}
\value{
一个数据框，包含所有 Excel 文件的 `planting` 表合并后的结果。
}
\description{
该函数会读取指定文件夹中的所有 Excel 文件的 `planting` 表，确保表中的字段顺序为
"id", "user", "stageid", "name", "ma", "pa", "mapa", "memo", "stage", "next_stage", "f"。
如果缺少某些字段，则补齐这些字段并填充 NA；如果有多余的字段，则删除这些字段。
}
