% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{read_df_from_sqlite}
\alias{read_df_from_sqlite}
\title{Read Data Frame from SQLite Database}
\usage{
read_df_from_sqlite(db_path, table_name)
}
\arguments{
\item{db_path}{A string specifying the path to the SQLite database file.}

\item{table_name}{A string specifying the name of the table to read the data from.}
}
\value{
A data frame containing the data read from the specified table.
}
\description{
This function reads data from a specified table in an SQLite database and returns it as a data frame.
}
\examples{
\dontrun{
df <- read_df_from_sqlite("my_database.sqlite", "my_table")
print(df)
}
}
