% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_plot.r
\name{replicate_rows}
\alias{replicate_rows}
\title{根据最后一列的值复制数据框的行}
\usage{
replicate_rows(df)
}
\arguments{
\item{df}{一个数据框，其中最后一列包含指示每行复制次数的整数值。}
}
\value{
一个根据最后一列的值复制行的新数据框。
}
\description{
该函数根据数据框最后一列中的整数值来复制行。每一行将根据对应的
最后一列的值重复相应的次数。
}
\examples{
df <- data.frame(
  name = c("A", "B", "C"),
  value = c(10, 20, 30),
  times = c(1, 3, 2)
)
replicate_rows(df)

}
