% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.R
\name{summarize_data}
\alias{summarize_data}
\title{对大豆数据进行汇总}
\usage{
summarize_data(updated, soy_traits)
}
\arguments{
\item{updated}{数据框，包含原始大豆数据}

\item{soy_traits}{数据框，包含大豆性状信息，用于指定需要汇总的数值型列}
}
\value{
返回一个数据框，包含汇总后的数据
}
\description{
该函数用于对大豆数据进行汇总，包括：
 * 筛选出重复次数为1的数据
 * 去除需要汇总的数值型列
 * 对数值型列进行分组汇总，计算平均值
 * 将汇总结果合并回原始数据
}
