% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baiaoyun.R
\name{update_baiaoyun_old}
\alias{update_baiaoyun_old}
\title{更新百奥云数据集中的性状信息}
\usage{
update_baiaoyun_old(df, updf, overwrite = TRUE)
}
\arguments{
\item{df}{主数据集，包含需要更新的性状信息。必须包含唯一编号列和性状列。}

\item{updf}{更新数据集，包含需要更新的性状信息以及唯一编号。列名应包含字段ID和性状信息。}

\item{overwrite}{逻辑值，是否覆盖主数据集中数据，TRUE覆盖，FALSE不覆盖。}
}
\value{
更新后的主数据集 `df`。
}
\description{
根据提供的更新数据集 `updf` 更新主数据集 `df` 中的性状信息。
}
\details{
函数的工作流程如下：
\itemize{
  \item 找到主数据集中唯一编号的列。
  \item 遍历更新数据集中的每个性状，找到对应的中文名称，并在主数据集中找到对应列。
  \item 根据唯一编号，将更新数据集中的性状信息更新到主数据集中。
}
}
\examples{
df <- data.frame("唯一编号" = c(1, 2, 3), "性状1" = c(NA, 2, 3))
updf <- data.frame("fieldid" = c(1, 2), "性状1" = c(10, NA))
updated_df <- update_baiaoyun(df, updf)
print(updated_df)
}
