% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{update_table_from_df}
\alias{update_table_from_df}
\title{从数据框更新SQLite数据库表}
\usage{
update_table_from_df(db_path, table_name, data_df)
}
\arguments{
\item{db_path}{字符串，SQLite数据库的文件路径。}

\item{table_name}{字符串，要更新的表名。}

\item{data_df}{数据框，包含要更新的数据。数据框的第一列必须是fieldid。}
}
\value{
无返回值。如果更新成功，则提交事务；如果发生错误，则回滚事务。
}
\description{
该函数用于从数据框更新SQLite数据库表。
}
\examples{
\dontrun{
db_path <- "path/to/database.db"
table_name <- "my_table"
data_df <- data.frame(fieldid = 1:3, value = c("A", "B", "C"))
update_table_from_df(db_path, table_name, data_df)
}
}
