% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{write_df_to_sqlite}
\alias{write_df_to_sqlite}
\title{Write Data Frame to SQLite Database}
\usage{
write_df_to_sqlite(df, db_path, table_name, overwrite = FALSE, append = FALSE)
}
\arguments{
\item{df}{A data frame to be written to the SQLite database.}

\item{db_path}{A string specifying the path to the SQLite database file.}

\item{table_name}{A string specifying the name of the table to write the data to.}

\item{overwrite}{A logical value indicating whether to overwrite the table if it exists. Default is FALSE.}

\item{append}{A logical value indicating whether to append the data to the table if it exists. Default is FALSE.}
}
\description{
This function writes a data frame to a specified table in an SQLite database.
}
\examples{
\dontrun{
df <- data.frame(id = 1:3, name = c("Alice", "Bob", "Charlie"))
write_df_to_sqlite(df, "my_database.sqlite", "my_table", overwrite = TRUE)
}
}
