% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_model_excel}
\alias{write_model_excel}
\title{将模型数据写入 Excel 文件}
\usage{
write_model_excel(data, file_name = "soyplant_model.xlsx", dir_path = getwd())
}
\arguments{
\item{data}{列表，由 `generate_model_data()` 生成的数据列表。}

\item{file_name}{字符串，导出的 Excel 文件名。默认为 `"soyplant_model.xlsx"`。}

\item{dir_path}{字符串，保存文件的目标目录。默认为当前工作目录。}
}
\value{
字符型，文件保存路径。
}
\description{
将模型数据写入 Excel 文件
}
