% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{download_model}
\alias{download_model}
\title{下载模型数据并导出为 Excel 文件}
\usage{
download_model(file_name = "soyplant_model.xlsx", dir_path = getwd())
}
\arguments{
\item{file_name}{字符串，导出的 Excel 文件名。默认为 `"soyplant_model.xlsx"`。}

\item{dir_path}{字符串，保存文件的目标目录。默认为当前工作目录。}
}
\value{
无返回值。函数执行后将在指定目录生成 Excel 文件。
}
\description{
此函数依次从 mapa 中提取组合、群体、单株、系、初选、高代和种植数据，
并将这些数据写入 Excel 文件。支持自定义文件名与保存目录。
}
