% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_id.R
\name{generate_id}
\alias{generate_id}
\title{根据指定参数生成ID序列}
\usage{
generate_id(
  start_num = 1,
  end_num = 10,
  char = "g",
  digit_length = 5,
  include_datetime = TRUE
)
}
\arguments{
\item{start_num}{整数，表示ID序列的起始号码。}

\item{end_num}{整数，表示ID序列的结束号码。}

\item{char}{字符，放在每个ID的开头。}

\item{digit_length}{整数，指定数字部分的长度，控制填充。}

\item{include_datetime}{逻辑值，是否在ID中包含当前的年份、月份、日期、小时、分钟和秒。}
}
\value{
一个包含所有生成ID的向量，从start_num到end_num。
}
\description{
该函数根据指定的参数生成一系列ID。
}
\examples{
\dontrun{
start <- 123
end <- 135
character <- "A"
digit_length <- 5
result_ids <- generate_id(start, end, character, digit_length, include_datetime = TRUE)
for (result_id in result_ids) {
  print(paste("生成的ID:", result_id))
}
}

}
