% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination.R
\name{get_combination}
\alias{get_combination}
\title{获得组合列表}
\usage{
get_combination(
  filename,
  prefix = "ZJ",
  startN = 1,
  only = TRUE,
  order = FALSE
)
}
\arguments{
\item{filename}{文件名，里面包括3列，顺序为为母本，父本和备注，其它列无要求，文件格式为xlsx}

\item{prefix}{组合前缀}

\item{startN}{起始编号}

\item{only}{逻辑值 是否去重}

\item{order}{逻辑值，是否按亲本排序}
}
\value{
返回杂交表
}
\description{
根据父母本获得组合列表
}
