% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_productivity.R
\name{get_primary}
\alias{get_primary}
\title{处理株行进入初级产比的数据}
\usage{
get_primary(my_line, start_num = 1)
}
\arguments{
\item{my_line}{数据框，包含株行信息}

\item{start_num}{整数，起始编号，默认为1}
}
\value{
数据框，处理后的初级产比数据
}
\description{
处理株行进入初级产比的数据
}
