% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planting.R
\name{planting}
\alias{planting}
\title{计划种植试验}
\usage{
planting(
  my_primary,
  ck = c("冀豆12", "冀豆17"),
  interval = 3,
  s_prefix = "GC",
  rp = 2,
  treatment = c(""),
  place = c("石家庄", "德州"),
  ckfixed = TRUE,
  digits = 3,
  rows = 6
)
}
\arguments{
\item{my_primary}{数据框，数据框中要包含id,stageid,name三列}

\item{ck}{向量，对照名称}

\item{interval}{插入对照的材料间隔数}

\item{s_prefix}{材料前缀}

\item{rp}{种植材料重复数，第一重复顺序，其它重复随机，并排两列材料不相同}

\item{treatment}{向量，处理，可以多个水平}

\item{place}{向量，试验地点，可多个试验地点}

\item{ckfixed}{逻辑值，对照是否固定，是固定则按等材料数插入}
}
\value{
返回插入对照和重复的数据框，给材料进行了编号，对照随机
}
\description{
计划种植试验
}
