% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savewb.R
\name{savewb}
\alias{savewb}
\title{保存数据表}
\usage{
savewb(origin = NA, planting = NA, myview = NA, filename, overwrite = FALSE)
}
\arguments{
\item{origin}{数据框，保存生成的原始数据}

\item{planting}{数据框，保存生成的种植数据}

\item{myview}{数据框，保存调查表}

\item{filename}{字符串，保存的文件名}

\item{overwrite}{是否覆盖}
}
\value{
保存生成的workbook类型数据到excel中
}
\description{
保存数据表
}
