% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{traitstable}
\alias{traitstable}
\title{生成性状表格}
\usage{
traitstable(trait_col = 3, validation_number = 100)
}
\arguments{
\item{trait_col}{数字，性状开始的列}

\item{validataion_number}{数字，验证的行数}
}
\value{
返回openxlsx包createWorkbook()创建对象
}
\description{
返回openxlsx包createWorkbook()创建对象，包括两个表，性状分类表和带验证功能的性状表
}
